Revert "Feat/signalert (#46)"
This reverts commit e34ea5eb
