+++ .env.sample
@@ -12,7 +12,6 @@ GLD_ENIGMA=761176450564096000
 ROLE_RETAIL=1161967553476825088
 ROLE_CYBER=1161967520845144135
 # DISCORD CHANNELS
-CHNL_SIGN=1168538353268957184
 CHNL_RETAIL=1161995902060331019
 CHNL_RETAIL_ALERT=1165639947349397544
 CHNL_CYBER=1161995870103949324
