--- src/models/Mail.ts
@@ -1,11 +1,0 @@
-export default interface Mail {
-  createdDateTime: string;
-  subject: string;
-  bodyPreview: string;
-  sender: {
-    emailAddress: {
-      name: string;
-      address: string;
-    }
-  }
-}
