--- src/models/MailServiceData.ts
@@ -1,5 +1,0 @@
-export default interface MailServiceData {
-    date: number;
-    morning: boolean;
-    afternoon: boolean;
-}
